One Front Door/SourceCode/MerchAngular4/dpsg.merchandising.web/src/app/planning/rsroute/rsroute.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { RsrouteComponent } from './rsroute.component';

describe('Component: Rsroute', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [RsrouteComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([RsrouteComponent],
      (component: RsrouteComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(RsrouteComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(RsrouteComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-rsroute></app-rsroute>
  `,
  directives: [RsrouteComponent]
})
class RsrouteComponentTestController {
}

